requirements.txt
python-dotenv==1.0.1
loguru==0.7.3
SQLAlchemy==2.0.40
aiosqlite==0.21.0
greenlet==3.0.3
web3==6.20.4
fake-useragent==2.2.0
aiohttp-socks==0.8.4
curl_cffi==0.11.3
aiohttp==3.11.8
cryptography==44.0.2
aiohttp_proxy==0.1.2
py-solc-x==2.0.3
inquirer~=3.4.0
rich~=13.9.4
colorama~=0.4.6
Faker~=30.4